11-08

+++ src/main/java/com/zlsoft/service/impl/TvDatasourceFormatServiceImpl.java
@@ -1,17 +1,22 @@
 package com.zlsoft.service.impl;
 
+import com.zlsoft.core.Result;
 import com.zlsoft.dao.TvDatasourceDataMapper;
 import com.zlsoft.dao.TvDatasourceFormatMapper;
 import com.zlsoft.model.entity.TvDatasourceData;
 import com.zlsoft.model.entity.TvDatasourceFormat;
 import com.zlsoft.service.TvDatasourceFormatService;
 import com.zlsoft.core.AbstractService;
+import lombok.extern.log4j.Log4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.RestTemplate;
 import tk.mybatis.mapper.entity.Condition;
 
 import javax.annotation.Resource;
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -21,6 +26,7 @@ import java.util.Map;
  * Created by DP on 2019/07/25.
  */
 @Service
+
 @Transactional
 public class TvDatasourceFormatServiceImpl extends AbstractService<TvDatasourceFormat> implements TvDatasourceFormatService {
     @Resource
@@ -28,6 +34,11 @@ public class TvDatasourceFormatServiceImpl extends AbstractService<TvDatasourceF
 
     @Autowired
     private TvDatasourceDataMapper tvDatasourceDataMapper;
+
+    @Autowired
+    private RestTemplate restTemplate;
+    @Value("${api.url}")
+    private String URL;
 
     @Override
     public void timingUpdate() {
@@ -45,21 +56,35 @@ public class TvDatasourceFormatServiceImpl extends AbstractService<TvDatasourceF
 
             }
             // 根据SQL获取跟新数据
-            if (tvDatasource.getType() == 2) {
+            /*if (tvDatasource.getType() == 2) {
                 List<Map<String, Object>> updateBySQL = tvDatasourceFormatMapper.getUpdateBySQL(tvDatasource.getQuerySql());
                 if (updateBySQL.size() > 0) {
                     insertAll.addAll(updateBySQL);
                 }
                 //tvDatasourceMapper.updateByPrimaryKeySelective(datasource);
+            }*/
+
+            // 根据SQL到his获取shuju
+            if (3 == tvDatasource.getType()) {
+                Result result = restTemplate.getForObject(URL + "/tv/bySql?sql=" + tvDatasource.getQuerySql() + "&format_id=" + tvDatasource.getId() + "&data_name=" + tvDatasource.getDataName() + "&abbreviation=" + tvDatasource.getAbbreviation(), Result.class);
+                if (null != result) {
+                    List<Map<String,Object>> data = (List<Map<String,Object>>) result.getData();
+                    if (data.size() > 0) {
+                        insertAll.addAll(data);
+                    }
+                }
             }
 
         }
         // 删除所有表格的数据
-        Condition condition = new Condition(TvDatasourceData.class);
-        condition.createCriteria().andCondition("1=1");
-        tvDatasourceDataMapper.deleteByCondition(condition);
-        // 开始增加
-        tvDatasourceDataMapper.insertByMap(insertAll);
+        if (insertAll.size() > 0) {
+            Condition condition = new Condition(TvDatasourceData.class);
+            condition.createCriteria().andCondition("1=1");
+            tvDatasourceDataMapper.deleteByCondition(condition);
+            // 开始增加
+            tvDatasourceDataMapper.insertByMap(insertAll);
+        }
+
 
     }
